requirements.txt
joblib
matplotlib
numpy
progressbar2
qutip
